Fix prisma generate load time (#168)
Fixes some of #151.
I've checked the DockerFile.dev to make the `npx prisma generate`
faster.

Tried a few things and what worked was to remove the **installer** step,
it worked perfectly with build times around 60 seconds(image downloads
included) on my machine. Re cloned the repo in an empty directory and
pruned all docker stuff and retested, same result.

Dived a bit into this with the docker tool called
[Dive](https://github.com/wagoodman/dive). When I inspected the layers
of the docker image, I found that the only files that maybe got lost
between (from) steps were some cache stuff about NPM in **/TMP** and
**/ROOT**.


![Screenshot_1](https://user-images.githubusercontent.com/94180156/204118613-f48ad019-43a1-4ec5-bd7d-817c48fff018.png)


Perhaps a bug in Prisma, maybe he needs the cache files (because there
was some in **/ROOT**), I don't know… :(


Also, I tried the live reload stuff that **remix dev** command does, so
I changed this:
```javascript
<GappedBox css={{ alignItems: "center" }}>pano</GappedBox>
```
To this
```javascript
<GappedBox css={{ alignItems: "center" }}>panoa</GappedBox>
```

And it did not auto reload, don't know if it was the case before because
mine never autoreloaded. Found an issue on this not working in remix :
[[Bug]: Live Reload not
working.](https://github.com/remix-run/remix/issues/1601#top)