auto reload for macos and windows (#187)
fixes #184 

Remix uses a websocket for development server, in our configs it uses
the port 8002. So adding this port to the docker compose makes auto
reload work for macOS and windows.

For windows, a WSL extension is needed and the repository files must be
in WSL filesystem. In /home/username or /root.
The extension crates a vscode server in WSL and connects your VSCode
into this server.

You can also click here to initiate the connexion manually, if not
setup, it should setup the server automatically.

![image](https://user-images.githubusercontent.com/94180156/208272422-b55d44a7-1714-461b-b290-018b196cec97.png)

Co-authored-by: c.soylu <c.soylu@geco-it.fr>