feat(@kampus-apps/pano): add settings page validations using zod (#226)
### Defect Fixes

This PR introduces `zod` for form validations. It starts with
`settings.tsx` page where user can change username, email and password.
This is the first implementation, so writing a wrapper around this could
be great.

Maybe we should move these zod schemas somewhere out of this page so we
won't import `zod` here and we can just import schemas to validate.