Sort posts (#192)
Closes #71 

This pull request adds the following functionality:

- 3 new routes under the post route, hot, active, and liked
- A new PostSorter component
- Sorting operations

## Main concerns:
- I was going to introduce another `<Outlet />` for these routes in
order to have only one `<PostSorter />` component. When I create an
index route for posts and have outlet for these routes, it affects other
pages such as "$slug-$id" route which is not what we want. For that
reason, I had to render `<PostSorter />` at each route. Since PostSorter
acts accordingly to url, there shouldn't be a state problem but I would
like to hear your suggestions and come up with a better solution.
- PostSorter component's placement & styling

## Screenshots

![image](https://user-images.githubusercontent.com/8138047/208287835-f8f67027-891e-4977-a2f6-28539cd914a9.png)

![image](https://user-images.githubusercontent.com/8138047/208287733-f0b57af5-81eb-4f69-920e-0437db4a06a5.png)

![image](https://user-images.githubusercontent.com/8138047/208287801-461c9734-38e6-4244-b055-d9c4db1836ac.png)

![image](https://user-images.githubusercontent.com/8138047/208287821-212a72a0-7b10-419f-bf1d-98716fc54464.png)