feat(apps/pano): Update username generation from email logic, add ran… (#322)
…dom suffix for conflicts

# Description

A prototype that fixes #318 

- Updates to username generation logic (f.e: `berke@kamp.us` would
translate to `berke`)
- Random suffix addition for username conflicts (f.e: `berke@kamp.us`
would translate to `berke4323` if a user with username `berke` already
exists)

### Checklist

- [X] discord username: `bumshaqalaqa#2973`
- [X] Closes #318
- [X] PR must be created for an issue from issues under "In progress"
column from [our project
board](https://github.com/orgs/kamp-us/projects/2/views/1).
- [X] A descriptive and understandable title: The PR title should
clearly describe the nature and purpose of the changes. The PR title
should be the first thing displayed when the PR is opened. And it should
follow the semantic commit rules, and should include the
app/package/service name in the title. For example, a title like
"docs(@kampus-apps/pano): Add README.md" can be used.
- [X] Related file selection: Only relevant files should be touched and
no other files should be affected.
- [X] I ran `npx turbo run` at the root of the repository, and build was
successful.
- [X] I installed the npm packages using `npm install --save-exact
<package>` so my package is pinned to a specific npm version. Leave
empty if no package was installed. Leave empty if no package was
installed with this PR.

### How were these changes tested?

No testing was made for these changes.