feat(@kampus/pano-button & @kampus/pano): Create an embeddable share component (#345)
# Description

I have created a new package containing a share button, that could be
embedded in websites to point out to send route of pano with a prefilled
URL value.
I have also added a query parameter to pano "/send" route, that is being
checked for prefill the URL and title fields

### Checklist

- [X] discord username: `bumshaqalaqa#2973`
- [X] Closes #310 
- [X] PR must be created for an issue from issues under "In progress"
column from [our project
board](https://github.com/orgs/kamp-us/projects/2/views/1).
- [X] A descriptive and understandable title: The PR title should
clearly describe the nature and purpose of the changes. The PR title
should be the first thing displayed when the PR is opened. And it should
follow the semantic commit rules, and should include the
app/package/service name in the title. For example, a title like
"docs(@kampus-apps/pano): Add README.md" can be used.
- [X] Related file selection: Only relevant files should be touched and
no other files should be affected.
- [X] I ran `npx turbo run` at the root of the repository, and build was
successful.
- [X] I installed the npm packages using `npm install --save-exact
<package>` so my package is pinned to a specific npm version. Leave
empty if no package was installed. Leave empty if no package was
installed with this PR.

### How were these changes tested?

I have added a temporary share button to the settings page, imported it
from @kampus/pano-button package, and provided the component a URL
value, then tested it manually.