feat(apps/pano): Notification service (#365)
# Description

First iteration of the notification service for Pano, could be improved
in further iterations.

### Functionalities:
- User gets a notification for an event (such as comment/upvote), on a
post/comment that they own.
- An unread identifier will appear on top-right of the notifications
icon if there are unread notifications.
- Notifications could be marked as read by clicking on them or clicking
on "Hepsini okundu olarak işaretle" button.
- Unread identifier will disappear if all notifications are marked as
read.

### There are 4 types of notifications: 
- "Upvote on comment": When user gets an upvote on their comment, they
will get a notification linking to the comment which got upvoted
alongside the username who upvoted the comment.
- "Upvote on post": When user gets an upvote on their post, they will
get a notification linking to the post which got upvoted alongside the
username who upvoted the post.
- "Comment on post": When user gets a comment on their post, they will
get a notification linking to the comment under their post alongside the
writer's username.
- "Reply to comment": When a user gets a reply to their comment, they
will get a notification linking to the comment **that got replied**,
alongside the replier's username.

### Constraints:
- If a user gets a reply to their comment under their own post, they
will only get the notification for the reply.
- A user can not get notifications for their actions on their own
posts/comments.
- Notification service could show only last 10 notifications for now.

### Example Scenario:
1. Let's have two actors with usernames "berke" and "anday", let's say
"anday" creates a post and a comment,

![flow-1](https://user-images.githubusercontent.com/31995815/230725618-d70dc607-a49e-4af0-872a-df26693ebc09.png)

2. "berke" logs in and likes both the post and the comment.

3. "anday" is now able to see new notifications appear on the top right.

![flow-2](https://user-images.githubusercontent.com/31995815/230725675-1de69e8a-88b8-4c43-9378-b421ae9d7e35.png)

4. "anday" clicks to one of the notifications.

![flow-3](https://user-images.githubusercontent.com/31995815/230725687-f6aa4274-abaf-4ad3-9e6d-a17ea14c5ae5.png)

5. The link takes "anday" to the relevant comment, marking it.

![flow-4](https://user-images.githubusercontent.com/31995815/230725695-f4d5c91d-bf8d-41fc-9678-46d9e8fa8a2e.png)

6. For some reason "anday" deletes the post, all the notifications
related to it gets deleted.

![flow-5](https://user-images.githubusercontent.com/31995815/230725864-eb0f1464-40c1-4cfb-9a83-e3ab619ba57b.png)


### Checklist

- [X] discord username: `bumshaqalaqa#2973`
- [X] Closes #360 
- [X] PR must be created for an issue from issues under "In progress"
column from [our project
board](https://github.com/orgs/kamp-us/projects/2/views/1).
- [X] A descriptive and understandable title: The PR title should
clearly describe the nature and purpose of the changes. The PR title
should be the first thing displayed when the PR is opened. And it should
follow the semantic commit rules, and should include the
app/package/service name in the title. For example, a title like
"docs(@kampus-apps/pano): Add README.md" can be used.
- [X] Related file selection: Only relevant files should be touched and
no other files should be affected.
- [X] I ran `npx turbo run` at the root of the repository, and build was
successful.
- [X] I installed the npm packages using `npm install --save-exact
<package>` so my package is pinned to a specific npm version. Leave
empty if no package was installed. Leave empty if no package was
installed with this PR.

### How were these changes tested?

Changes tested manually on different user scenarios.

---------

Co-authored-by: Umut Sirin <umutsirin1@gmail.com>