feat(@kampus/tailwind): create defaultTheme (#464)
nothing important, i made this thinking that i will use the tailwind
theme in #456, but i ended up using the preset instead. But still wanted
to keep this as this cleans up the code pretty nicely, and also exported
it at package level in case we want to use it somewhere else.