feat(pano): add create post form with modal route ux (#638)
Add modal based UX for pano post creation. Uses [parallel routes] and
"[intercepting routes]" patterns from next.js to render a modal if we go
to `/post/create` route from application using a link, and render the
form directly if user lands on that route directly by either clicking to
that link from somewhere else, or reloading at that route.

_(an example implementation from next.js:
https://github.com/vercel-labs/nextgram)_


https://github.com/kamp-us/monorepo/assets/1783869/e56072c3-a938-4cc6-b228-bee5d11d9349

### Checklist

- [ ] discord username: `umut`
- [ ] Closes #613 
- [ ] A descriptive and understandable title: The PR title should
clearly describe the nature and purpose of the changes. The PR title
should be the first thing displayed when the PR is opened. And it should
follow the semantic commit rules, and should include the
app/package/service name in the title. For example, a title like
"docs(@kampus-apps/pano): Add README.md" can be used.
- [ ] Related file selection: Only relevant files should be touched and
no other files should be affected.
- [ ] I ran `npx turbo run` at the root of the repository, and build was
successful.
- [ ] I installed the npm packages using `npm install --save-exact
<package>` so my package is pinned to a specific npm version. Leave
empty if no package was installed. Leave empty if no package was
installed with this PR.

[parallel routes]:
https://nextjs.org/docs/app/building-your-application/routing/parallel-routes
[intercepting routes]:
https://nextjs.org/docs/app/building-your-application/routing/intercepting-routes